requirements.txt
disnake~=2.9.1
pytz~=2024.2
DateTime~=5.3
python-dotenv~=1.0.0